refactor toast logic
