+++ src/features/toast/index.js
@@ -1,3 +1,3 @@
-export {default as ToastReducer} from './reducer';
+export {ToastReducer} from './reducer';
 export {useVisibleToasts, useToastInfo} from './selectors';
-export {useToastActions} from './actionCreators';
+export {useToastActions} from './useToastActions';
